Refactor testimonials management: Move testimonials data to a centralized data file for better organization and maintainability. Update CustomerSatisfactionBanner and TestimonialCard components to utilize the new data structure. Replace image tags with Next.js Image component for optimized loading and responsiveness.
