+++ components/ui/testimonial-card.tsx
@@ -19,7 +19,7 @@ export const TestimonialCard = ({ text, image, name, role }: Testimonial) => {
           width={40}
           height={40}
           loading="lazy"
-          src={image}
+          src={`/clients/${image}`}
           alt={name}
           className="h-10 w-10 rounded-full"
         />
